Input throttling and configurable feedback
If the glove firmware sends data more often than we poll for it then the buffer will become saturated and block future reads. Data is guaranteed to be ready on the next read so just purge the buffers after we get a full set of encoded data.

Added config to enable/disable feedback for gloves that do not support it.