Update max_analog_value default config & add check
Changes the "encoding" field in the default.vrsettings to
"encoding_alpha" and adds a check to device_configuration.cpp to catch
it if it's zero.

If this value is set to zero, a division by zero happens in alpha
encoding service that results in a NaN value, which cascades into
problems setting bone position in knuclke_device_driver.cpp.

This bug might manifest itself in a Unity game by an error message that
reads:

`transform.localPosition assign attempt for ... is not valid. Input
localPosition is { NaN, NaN, NaN }`